backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
fastapi==0.103.1
uvicorn==0.23.2
google_auth_oauthlib==1.0.0
google-api-python-client==2.100.0
python-multipart==0.0.6
httpx # Added for making async HTTP requests
google-generativeai # Added for Gemini API
google-auth # Add this line
